separate d3 component from mui card to fix rendering issue
